fix: username dropdown behavior on mobile (#2099)
This pull request introduces device detection logic to improve the
responsiveness of the account search dropdown. The most important
changes are grouped below:

**Device detection and responsive UI:**

* Added a new `useDevice` hook in `hooks/device.ts` to detect if the
user is on a mobile or desktop device, and whether the app is running as
a PWA. This hook uses window width and media queries.
* Updated `AccountSearchDropdown` to use the `useDevice` hook and adjust
the maximum number of search results shown: 3 for mobile, 5 for desktop.
[[1]](diffhunk://#diff-01cd99b06f0175e14ec291c83f567302e9f428f2d9507121e8decec9da70c46cR15)
[[2]](diffhunk://#diff-01cd99b06f0175e14ec291c83f567302e9f428f2d9507121e8decec9da70c46cR45-R59)
* Added an `isLoading` prop to `AccountSearchDropdown` and its parent
components to allow external control of loading state.
[[1]](diffhunk://#diff-01cd99b06f0175e14ec291c83f567302e9f428f2d9507121e8decec9da70c46cR25)
[[2]](diffhunk://#diff-01cd99b06f0175e14ec291c83f567302e9f428f2d9507121e8decec9da70c46cL69-R88)
[[3]](diffhunk://#diff-97aa791e2d2e668e97f77435d49884b01c5828e8b30a3045999464eec464378bR333)
[[4]](diffhunk://#diff-97aa791e2d2e668e97f77435d49884b01c5828e8b30a3045999464eec464378bR343)
* Improved UX by auto-selecting the first dropdown item if the input
query matches the first result.